% Generated by roxygen2 (4.0.1): do not edit by hand
\name{averageMSnSet}
\alias{averageMSnSet}
\title{Generate an average \code{MSnSet}}
\usage{
averageMSnSet(x, avg = function(x) mean(x, na.rm = TRUE), disp = function(x)
  mad(x, na.rm = TRUE))
}
\arguments{
\item{x}{A \code{list} of valid \code{MSnSet} instances to be averaged.}

\item{avg}{The averaging function. Default is hte median after
removing missing valiues, as computed by \code{function(x)
median(x, na.rm = TRUE)}.}

\item{disp}{The disperion function. Default is the median absolute
deviation as computed by \code{mad(x, na.rm = TRUE)}. See
\code{?mad} for details. Note that the \code{mad} of a single
value is 0 (as opposed to \code{NA} for the standard deviation,
see example below).}
}
\value{
A new average \code{MSnSet}.
}
\description{
Given a list of \code{MSnSet} instances, typically representing
replicated experiments, the function returns an average
\code{MSnSet}.
}
\details{
This function is aimed at facilitating the visualisation of
replicated experiments and should not be used as a replacement for
a statistical analysis.

The samples of the instances to be averaged must be identical but
can be in a different order (they will be reordered by
default). The features names of the result will correspond to the
union of the feature names of the input \code{MSnSet}
instances. Each average value will be computed by the \code{avg}
function and the dispersion of the replicated measurements will be
estimated by the \code{disp} function. These dispersions will be
stored as a \code{data.frame} in the feature metadata that can be
accessed with \code{fData(.)$disp}. Similarly, the number of
missing values that were present when average (and dispersion)
were computed are available in \code{fData(.)$disp}.
}
\examples{
library("pRolocdata")
## 3 replicates from Tan et al. 2009
data(tan2009r1)
data(tan2009r2)
data(tan2009r3)
avg <-averageMSnSet(list(tan2009r1, tan2009r2, tan2009r3))
dim(avg)
head(exprs(avg))
head(fData(avg)$nNA)
head(fData(avg)$disp)
## using the coefficient of variation as measure of dispersion
cv <- function(x, na.rm = TRUE) {
   sdx <- sd(x, na.rm = na.rm)
   sdx/abs(mean(x, na.rm = na.rm))
}
avg2 <-averageMSnSet(list(tan2009r1, tan2009r2, tan2009r3),
                     disp = cv)
head(fData(avg2)$disp)
## keep only complete observations, i.e proteins
## that had 0 missing values for all samples
sel <- apply(fData(avg2)$nNA, 1 , function(x) all(x == 0))
avg2 <- avg2[sel, ]
disp <- rowMax(fData(avg2)$disp)
setStockcol(paste0(getStockcol(), "AA"))
plot2D(avg, cex = 7.7 * disp)
title(main = paste("Dispersion: coefficient of variation",
                   paste(round(range(disp), 3), collapse = " - ")))
}
\author{
Laurent Gatto
}

