% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\docType{class}
\name{Backend-class}
\alias{Backend-class}
\title{Backend class}
\description{
Base class for all other [MSnExperiment-class] data
backend classes (e.g., [BackendMemory-class], [BackendHdf5-class]).
}
\details{
It is a *VIRTUAL* class and just defines a common interface for possible
backends.
}
\section{Implementation notes}{

An implementation *MUST* provide methods for the following generics:

- [backendReadSpectra()] to read spectra from the backend.
- [backendWriteSpectra()] to write spectra to the backend.

It may also provide methods for:

- [backendInitialize()] to setup the backend (create files, tables, ...).
- [backendImportData()] to initial import data from source *mzML* files.
}

\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend classes}
