% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendDeepCopy}
\alias{backendDeepCopy}
\title{Create a deep copy of the backend}
\usage{
backendDeepCopy(object, ..., BPPARAM = bpparam())
}
\arguments{
\item{object}{A [Backend-class] derivate.}

\item{...}{Other arguments passed to the methods.}

\item{BPPARAM}{Should parallel processing be used? See
[BiocParallel::bpparam()].}
}
\value{
A [Backend-class] derivate.
}
\description{
This generic is used to create a deep copy of the backend and its associated
files.
}
\details{
If an object with a file-based backend is copied in R via `newObj <- oldObj`
and subsequently modified the file content may be corrupted. The deep copy
also should copy all associated files to a new destination.

It should be only reimplemented if the backend uses files or other resources
that could be corrputed when the object is copied via `<-`, e.g. *HDF5*.
}
\seealso{
Other Backend generics: \code{\link{backendImportData}},
  \code{\link{backendInitialize}},
  \code{\link{backendReadSpectra}},
  \code{\link{backendWriteSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
