% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendImportData}
\alias{backendImportData}
\title{Import spectra data into a backend}
\usage{
backendImportData(object, files, ...)
}
\arguments{
\item{object}{An object inheriting from [Backend-class],
i.e. [BackendHdf5-class]}

\item{files}{The path to the source (generally .mzML) files.}

\item{...}{Other arguments passed to the methods.}
}
\value{
A [Backend-class] derivate.
}
\description{
This generic is used to import spectra data into a backend.
}
\details{
It should be only reimplemented if the backend is created from scratch
e.g. for *HDF5* the content of the .mzML files is imported into .h5 files.
It must not be reimplemented for the .mzML backend.
}
\seealso{
Other Backend generics: \code{\link{backendDeepCopy}},
  \code{\link{backendInitialize}},
  \code{\link{backendReadSpectra}},
  \code{\link{backendWriteSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
