% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendInitialize}
\alias{backendInitialize}
\title{Initialize a backend}
\usage{
backendInitialize(object, files, ...)
}
\arguments{
\item{object}{An object inheriting from [Backend-class],
i.e. [BackendHdf5-class]}

\item{files}{The path to the source (generally .mzML) files.}

\item{...}{Other arguments passed to the methods.}
}
\value{
A [Backend-class] derivate.
}
\description{
This generic is used to setup a backend.
}
\details{
It should be only reimplemented if the backend needs specific requirements,
e.g. for *HDF5* the creation of .h5 files; for *SQL* the creation of a
database or tables.
}
\seealso{
Other Backend generics: \code{\link{backendImportData}},
  \code{\link{backendReadSpectra}},
  \code{\link{backendWriteSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
