% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendMemory-class.R
\docType{class}
\name{BackendMemory-class}
\alias{BackendMemory-class}
\alias{BackendMemory}
\title{BackendMemory class}
\usage{
BackendMemory()
}
\value{
A [BackendMemory-class].
}
\description{
This class offers the *in-memory* reprensentation for the
[MSnExperiment-class] data. It mimics the classical [MSnExp-class] behaviour.
}
\section{Functions}{
\itemize{
\item \code{BackendMemory}: Constructor

This function is used to generated an *in-memory* backend. Just useful as
argument in [readMSnExperiment()].
}}

\section{Slots}{

\describe{
\item{\code{assay}}{A `list` containing the [Spectrum-class] objects.}
}}

\seealso{
Other Backend classes: \code{\link{Backend-class}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend classes}
