% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendReadSpectra}
\alias{backendReadSpectra}
\title{Read spectra data from a backend}
\usage{
backendReadSpectra(object, file, spectraData, ..., BPPARAM = bpparam())
}
\arguments{
\item{object}{An object inheriting from [Backend-class],
i.e. [BackendHdf5-class]}

\item{file}{The path to the source (generally .mzML) file.}

\item{spectraData}{A [S4Vectors::DataFrame-class]}

\item{...}{Other arguments passed to the methods.}

\item{BPPARAM}{Should parallel processing be used? See
[BiocParallel::bpparam()].}
}
\value{
A list of [Spectrum-class] objects.
}
\description{
This generic is used to read spectra data from a backend.
}
\details{
It *MUST* be reimplemented by all backends!
}
\seealso{
Other Backend generics: \code{\link{backendDeepCopy}},
  \code{\link{backendImportData}},
  \code{\link{backendInitialize}},
  \code{\link{backendWriteSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
