% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendWriteSpectra}
\alias{backendWriteSpectra}
\title{Write spectra data to a backend}
\usage{
backendWriteSpectra(object, file, spectra, spectraData, ...,
  BPPARAM = bpparam())
}
\arguments{
\item{object}{An object inheriting from [Backend-class],
i.e. [BackendHdf5-class]}

\item{file}{The path to the source (generally .mzML) file.}

\item{spectra}{A list of [Spectrum-class] objects that should be written to
the backend.}

\item{spectraData}{A [S4Vectors::DataFrame-class]}

\item{...}{Other arguments passed to the methods.}

\item{BPPARAM}{Should parallel processing be used? See
[BiocParallel::bpparam()].}
}
\value{
A [Backend-class] derivate.
}
\description{
This generic is used to write spectra data to a backend.
}
\details{
It *MUST* be reimplemented by all backends!
}
\seealso{
Other Backend generics: \code{\link{backendDeepCopy}},
  \code{\link{backendImportData}},
  \code{\link{backendInitialize}},
  \code{\link{backendReadSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
