% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R, R/functions-Chromatograms.R,
%   R/methods-Chromatograms.R
\docType{class}
\name{Chromatograms-class}
\alias{Chromatograms-class}
\alias{Chromatograms}
\alias{show,Chromatograms-method}
\title{Container for multiple Chromatogram objects}
\usage{
Chromatograms(data, ...)

\S4method{show}{Chromatograms}(object)
}
\arguments{
\item{data}{A \code{list} of \code{\link{Chromatogram}} objects.}

\item{...}{Additional parameters to be passed to the \code{\link{matrix}}
constructor, such as \code{nrow}, \code{ncol} and \code{byrow}.}
}
\description{
The \code{Chromatograms} class allows to store
    \code{\link{Chromatogram}} objects in a \code{matrix}-like
    two-dimensional structure.

\code{Chromatograms}: create an instance of class
    \code{Chromatograms}.
}
\details{
The \code{Chromatograms} class extends the base \code{matrix} class
    and hence allows to store \code{\link{Chromatogram}} objects in a
    two-dimensional array. Each row is supposed to contain
    \code{Chromatogram} objects for one MS data \code{slice} with a common
    mz and rt range. Columns contain \code{Chromatogram} objects from the
    same sample.
}
\seealso{
\code{\link{Chromatogram}} for the class representing chromatogram
    data.
    \code{\link{chromatogram}} for the method to extract a
    \code{Chromatograms} object from a \code{\linkS4class{MSnExp}} or
    \code{\linkS4class{OnDiskMSnExp}} object.
}
\author{
Johannes Rainer
}
