% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareMSnSets}
\alias{compareMSnSets}
\title{Compare two MSnSets}
\usage{
compareMSnSets(x, y, qual = FALSE, proc = FALSE)
}
\arguments{
\item{x}{First MSnSet}

\item{y}{Second MSnSet}

\item{qual}{Should the \code{qual} slots be compared? Default is
\code{FALSE}.}

\item{proc}{Should the \code{processingData} slots be compared?
Default is \code{FALSE}.}
}
\value{
A \code{logical}
}
\description{
Compares two \code{\linkS4class{MSnSet}} instances. The
\code{qual} and \code{processingData} slots are generally omitted.
}
\author{
Laurent Gatto
}
