\name{Deprecated}
\title{MSnbase Deprecated and Defunct}

\alias{MSnbase-deprecated}

%% \alias{readMzXML}

\description{
  The function, class, or data object you have asked for has been
  deprecated or made defunct. 

  Deprecated: \code{readMzXMLData}, \code{extractSpectra}.

  Defunct: none.
}

\keyword{internal}
