\name{dotproduct}
\alias{dotproduct}
\title{calculate the dot product between two vectors}
\usage{
dotproduct(x, y)
}
\arguments{
  \item{x}{double}

  \item{y}{double}
}
\value{
double, length == 1
}
\description{
Stein, S. E., and Scott, D. R. (1994). Optimization and
testing of mass spectral library search algorithms for
compound identification. Journal of the American Society
for Mass Spectrometry, 5(9), 859-866. doi:
http://dx.doi.org/10.1016/1044-0305(94)87009-8
}
\details{
Lam, H., Deutsch, E. W., Eddes, J. S., Eng, J. K., King,
N., Stein, S. E. and Aebersold, R. (2007) Development and
validation of a spectral library searching method for
peptide identification from MS/MS. Proteomics, 7: 655–667.
doi: http://dx.doi.org/10.1002/pmic.200600625
}

