% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MSnExp.R, R/utils.R
\docType{methods}
\name{extractMsData,MSnExp-method}
\alias{extractMsData,MSnExp-method}
\alias{extractMsData}
\alias{plotMsData}
\title{Extract MS data as a \code{data.frame} and visualize such MS data}
\usage{
\S4method{extractMsData}{MSnExp}(object, rt, mz, msLevel = 1L)

plotMsData(x, main = "", cex = 1, mfrow = c(2, 1),
  grid.color = "lightgrey", colramp = topo.colors)
}
\arguments{
\item{object}{A \code{MSnExp} (or \code{OnDiskMSnExp}) object.}

\item{rt}{\code{numeric(2)} with the retention time range from which the
data should be extracted.}

\item{mz}{\code{numeric(2)} with the m/z range.}

\item{msLevel}{\code{integer} defining the MS level(s) to which the data
should be sub-setted prior to extraction; defaults to
\code{msLevel = 1L}.}

\item{x}{\code{data.frame} such as returned by the \code{extractMsData} function.
Only a single \code{data.frame} should be submitted.}

\item{main}{\code{character(1)} specifying the title.}

\item{cex}{\code{numeric(1)} defining the size of points. Passed directly to the
\code{plot} function.}

\item{mfrow}{\code{numeric(2)} defining the plot layout. This will be passed
directly to \code{par(mfrow = mfrow)}. See \code{par} for more information.
Setting \code{mfrow = NULL} avoids calling \code{par(mfrow = mfrow)} hence
allowing to pre-define the plot layout.}

\item{grid.color}{a color definition for the grid line (or \code{NA} to skip
creating them).}

\item{colramp}{a \emph{color ramp palette} to be used to color the data points
based on their intensity. See argument \code{col.regions} in
\code{\link[lattice:level.colors]{lattice::level.colors()}} documentation.}
}
\value{
A \code{list} of length equal to the number of samples/files in
\code{object}. Each element being a \code{data.frame} with columns
\code{"rt"}, \code{"mz"} and \code{"i"} with the retention time, mz and
intensity tuples of a file. If no data is available for the mz-rt range
in a file a \code{data.frame} with 0 rows is returned for that file.
}
\description{
\code{extractMsData} extracts a \code{data.frame} of retention time, mz and
intensity values from each file/sample in the provided rt-m/z range
(or for the full data range if \code{rt} and \code{mz} are not defined).

\code{plotMsData} creates a plot that combines an (base peak) extracted
ion chromatogram on top (rt against intensity) and a plot of rt
against m/z values at the bottom.
}
\examples{
## Read some files from the test data package.
library(msdata)
library(MSnbase)
fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
raw_data <- readMSData(fls, mode = "onDisk")

## Read the full MS data for a defined mz-rt region.
res <- extractMsData(raw_data, mz = c(106.04, 106.06), rt = c(175, 187))

## We've got one data.frame per file
length(res)

## With number of rows:
nrow(res[[1]])

head(res[[1]])

## Plot the MS data for the first file.
plotMsData(res[[1]])
}
\seealso{
\link{MSnExp} for the data object.
}
\author{
Johannes Rainer
}
