% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factorsAsStrings}
\alias{factorsAsStrings}
\title{Converts factors to strings}
\usage{
factorsAsStrings(x)
}
\arguments{
\item{x}{A \code{data.frame}}
}
\value{
A \code{data.frame} where \code{factors} are converted to
\code{characters}.
}
\description{
This function produces the opposite as the \code{stringsAsFactors}
argument in the \code{data.frame} or \code{read.table} functions;
it converts \code{factors} columns to \code{characters}.
}
\examples{
data(iris)
str(iris)
str(factorsAsStrings(iris))
}
\author{
Laurent Gatto
}
