% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.amino.acids}
\alias{get.amino.acids}
\title{Amino acids}
\usage{
get.amino.acids()
}
\value{
A \code{data.frame}
}
\description{
Returns a \code{data.frame} of amino acid properties: \code{AA},
\code{ResidueMass}, \code{Abbrev3}, \code{ImmoniumIonMass},
\code{Name}, \code{Hydrophobicity}, \code{Hydrophilicity},
\code{SideChainMass}, \code{pK1}, \code{pK2} and \code{pI}.
}
\examples{
get.amino.acids()
}
\author{
Laurent Gatto
}
