% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.atomic.mass}
\alias{get.atomic.mass}
\title{Atomic mass.}
\usage{
get.atomic.mass()
}
\value{
A named \code{double}.
}
\description{
Returns a \code{double} of used atomic mass.
}
\examples{
get.atomic.mass()
}
\author{
Sebastian Gibb
}

