% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listOf}
\alias{listOf}
\title{Tests equality of list elements class}
\usage{
listOf(x, class)
}
\arguments{
\item{x}{A code{list}.}

\item{class}{A \code{character} defining the expected class.}
}
\value{
\code{TRUE} is all elements of \code{x} inherit from
\code{class}.
}
\description{
Compares equality of all members of a list.
}
\examples{
listOf(list(), "foo")
listOf(list("a", "b"), "character")
listOf(list("a", 1), "character")
}
\author{
Laurent Gatto
}

