% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCamelCase}
\alias{makeCamelCase}
\title{Convert to camel case by replacing dots by captial letters}
\usage{
makeCamelCase(x, prefix)
}
\arguments{
\item{x}{A \code{vector} to be transformed to camel case.}

\item{prefix}{An optional \code{character} of length one. Any
additional elements are ignores.}
}
\value{
A \code{character} of same length as \code{x}.
}
\description{
Convert a \code{vector} of characters to camel case by replacing
dots by captial letters.
}
\examples{
nms <- c("aa.foo", "ab.bar")
makeCamelCase(nms)
makeCamelCase(nms, prefix = "x")
}
\author{
Laurent Gatto
}
