% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R, R/functions-MChromatograms.R,
%   R/methods-MChromatograms.R
\docType{class}
\name{MChromatograms-class}
\alias{MChromatograms-class}
\alias{coerce,matrix,MChromatograms-method}
\alias{MChromatograms}
\alias{show,MChromatograms-method}
\alias{[,MChromatograms,ANY,ANY,ANY-method}
\alias{[<-,MChromatograms-method}
\alias{plot,MChromatograms,ANY-method}
\alias{phenoData,MChromatograms-method}
\alias{pData,MChromatograms-method}
\alias{pData<-,MChromatograms,data.frame-method}
\alias{$,MChromatograms-method}
\alias{$<-,MChromatograms-method}
\alias{colnames<-,MChromatograms-method}
\alias{sampleNames,MChromatograms-method}
\alias{sampleNames<-,MChromatograms,ANY-method}
\alias{isEmpty,MChromatograms-method}
\alias{featureNames,MChromatograms-method}
\alias{featureNames<-,MChromatograms-method}
\alias{featureData,MChromatograms-method}
\alias{featureData<-,MChromatograms,ANY-method}
\alias{fData,MChromatograms-method}
\alias{fData<-,MChromatograms,ANY-method}
\alias{fvarLabels,MChromatograms-method}
\alias{rownames<-,MChromatograms-method}
\alias{precursorMz,MChromatograms-method}
\alias{productMz,MChromatograms-method}
\alias{mz,MChromatograms-method}
\alias{polarity,MChromatograms-method}
\alias{bin,MChromatograms-method}
\alias{clean,MChromatograms-method}
\title{Container for multiple Chromatogram objects}
\usage{
MChromatograms(data, phenoData, featureData, ...)

\S4method{show}{MChromatograms}(object)

\S4method{[}{MChromatograms,ANY,ANY,ANY}(x, i, j, drop = FALSE)

\S4method{[}{MChromatograms}(x, i, j) <- value

\S4method{plot}{MChromatograms,ANY}(
  x,
  col = "#00000060",
  lty = 1,
  type = "l",
  xlab = "retention time",
  ylab = "intensity",
  main = NULL,
  ...
)

\S4method{phenoData}{MChromatograms}(object)

\S4method{pData}{MChromatograms}(object)

\S4method{pData}{MChromatograms,data.frame}(object) <- value

\S4method{$}{MChromatograms}(x, name)

\S4method{$}{MChromatograms}(x, name) <- value

\S4method{colnames}{MChromatograms}(x) <- value

\S4method{sampleNames}{MChromatograms}(object)

\S4method{sampleNames}{MChromatograms,ANY}(object) <- value

\S4method{isEmpty}{MChromatograms}(x)

\S4method{featureNames}{MChromatograms}(object)

\S4method{featureNames}{MChromatograms}(object) <- value

\S4method{featureData}{MChromatograms}(object)

\S4method{featureData}{MChromatograms,ANY}(object) <- value

\S4method{fData}{MChromatograms}(object)

\S4method{fData}{MChromatograms,ANY}(object) <- value

\S4method{fvarLabels}{MChromatograms}(object)

\S4method{rownames}{MChromatograms}(x) <- value

\S4method{precursorMz}{MChromatograms}(object)

\S4method{productMz}{MChromatograms}(object)

\S4method{mz}{MChromatograms}(object)

\S4method{polarity}{MChromatograms}(object)

\S4method{bin}{MChromatograms}(object, binSize = 0.5, breaks = numeric(), fun = max)

\S4method{clean}{MChromatograms}(object, all = FALSE, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{list} of \code{\link{Chromatogram}} objects.}

\item{phenoData}{either a \code{data.frame}, \code{AnnotatedDataFrame} or
\code{AnnotatedDataFrame} describing the phenotypical information of the
samples.}

\item{featureData}{either a \code{data.frame} or \code{AnnotatedDataFrame}
with additional information for each row of chromatograms.}

\item{...}{Additional parameters to be passed to the
\code{\link[base]{matrix}} constructor, such as \code{nrow}, \code{ncol}
and \code{byrow}.}

\item{object}{a \code{MChromatograms} object.}

\item{x}{For all methods: a \code{MChromatograms} object.}

\item{i}{For \code{[}: \code{numeric}, \code{logical} or \code{character}
defining which row(s) to extract.}

\item{j}{For \code{[}: \code{numeric}, \code{logical} or \code{character}
defining which columns(s) to extract.}

\item{drop}{For \code{[}: \code{logical(1)} whether to drop the
dimensionality of the returned object (if possible). The default is
\code{drop = FALSE}, i.e. each subsetting returns a \code{MChromatograms}
object (or a \code{Chromatogram} object if a single element is
extracted).}

\item{value}{For \code{[<-}: the replacement object(s). Can be a \code{list}
    of \code{\link{Chromatogram}} objects or, if length of \code{i} and
    \code{j} are 1, a single \code{\link{Chromatogram}} object.

    For \code{pData<-}: a \code{data.frame} with the number of rows matching
    the number of columns of \code{object}.

    For \code{colnames}: a \code{character} with the new column names.}

\item{col}{For \code{plot}: the color to be used for plotting. Either a
vector of length 1 or equal to \code{ncol(x)}.}

\item{lty}{For \code{plot}: the line type (see \code{plot} in the
\code{graphics} package
for more details. Can be either a vector of length 1 or of length equal
to \code{ncol(x)}.}

\item{type}{For \code{plot}: the type of plot (see
\code{plot} from the \code{graphics} package for more details.
Can be either a vector
of length 1 or of length equal to \code{ncol(x)}.}

\item{xlab}{for \code{plot}: the x-axis label.}

\item{ylab}{for \code{plot}: the y-axis label.}

\item{main}{for \code{plot}: the plot title. If not provided the mz range
will be used as plot title.}

\item{name}{For \code{$}, the name of the pheno data column.}

\item{binSize}{for \code{bin}: \code{numeric(1)} with the size of the bins
(in seconds).}

\item{breaks}{for \code{bin}: \code{numeric} defining the bins. Usually not
required as the function calculates the bins automatically based on
\code{binSize} and the retention time range of chromatograms in the same
row.}

\item{fun}{for \code{bin}: function to be used to aggregate the intensity
values falling within each bin.}

\item{all}{for \code{clean}: \code{logical(1)} whether all 0 intensities
should be removed (\code{all = TRUE}), or whether 0-intensities
adjacent to peaks should be kept (\code{all = FALSE}; default).}

\item{na.rm}{for \code{clean}: \code{logical(1)} whether all \code{NA}
intensities should be removed prior to clean 0 intensity data points.}
}
\value{
For \code{[}: the subset of the \code{MChromatograms} object. If a
    single element is extracted (e.g. if \code{i} and \code{j} are of length
    1) a \code{\link{Chromatogram}} object is returned. Otherwise (if
    \code{drop = FALSE}, the default, is specified) a \code{MChromatograms}
    object is returned. If \code{drop = TRUE} is specified, the method
    returns a \code{list} of \code{Chromatogram} objects.

    For \code{phenoData}: an \code{AnnotatedDataFrame} representing the
    pheno data of the object.

    For \code{pData}: a \code{data.frame} representing the pheno data of
    the object.

    For \code{$}: the value of the corresponding column in the pheno data
    table of the object.
}
\description{
The \code{MChromatograms} class allows to store
    \code{\link{Chromatogram}} objects in a \code{matrix}-like
    two-dimensional structure.

\code{MChromatograms}: create an instance of class
    \code{MChromatograms}.

\code{MChromatograms} objects can, just like a \code{matrix},
    be subsetted using the \code{[} method. Single elements, rows or columns
    can be replaced using e.g. \code{x[1, 1] <- value} where \code{value}
    has to be a \code{Chromatogram} object or a \code{list} of such objects.

\code{plot}: plots a \code{MChromatograms} object. For each row
    in the object one plot is created, i.e. all \code{\link{Chromatogram}}
    objects in the same row are added to the same plot.

\code{phenoData}: accesses the phenotypical desccription of the
    samples. Returns an \code{AnnotatedDataFrame} object.

\code{pData}: accesses the phenotypical description of the
    samples. Returns a \code{data.frame}.

\code{pData<-}: replace the phenotype data.

\code{$} and \code{$<-}: get or replace individual columns of
    the object's pheno data.

\code{colnames<-}: replace or set the column names of the
    \code{MChromatograms} object. Does also set the \code{rownames} of the
    \code{phenoData}.

\code{sampleNames}: get the sample names.

\code{sampleNames<-}: replace or set the sample names of the
    \code{MChromatograms} object (i.e. the \code{rownames} of the pheno data
    and \code{colnames} of the data matrix.

\code{isEmpty}: returns \code{TRUE} if the \code{MChromatograms}
    object or all of its \code{Chromatogram} objects is/are empty or contain
    only \code{NA} intensities.

\code{featureNames}: returns the feature names of the
    \code{MChromatograms} object.

\code{featureNames<-}: set the feature names.

\code{featureData}: return the feature data.

\code{featureData<-}: replace the object's feature data.

\code{fData}: return the feature data as a \code{data.frame}.

\code{fData<-}: replace the object's feature data by passing a
    \code{data.frame}

\code{fvarLabels}: return the feature data variable names (i.e.
    column names).

\code{rownames<-}: replace the rownames (and featureNames) of
    the object.

\code{precursorMz}: return the precursor m/z from the chromatograms. The
method returns a \code{matrix} with 2 columns (\code{"mzmin"} and
\code{"mzmax"}) and as many rows as there are rows in the
\code{MChromatograms} object. Each row contains the precursor m/z of the
chromatograms in that row. An error is thrown if the chromatograms within one
row have different precursor m/z values.

\code{productMz}: return the product m/z from the chromatograms. The
method returns a \code{matrix} with 2 columns (\code{"mzmin"} and
\code{"mzmax"}) and as many rows as there are rows in the
\code{MChromatograms} object. Each row contains the product m/z of the
chromatograms in that row. An error is thrown if the chromatograms within one
row have different product m/z values.

\code{mz}: returns the m/z for each row of the \code{MChromatograms} object
as a two-column \code{matrix} (with columns \code{"mzmin"} and
\code{"mzmax"}).

\code{polarity}: returns the polarity of the scans/chromatograms: `1`,
`0` or `-1` for positive, negative or unknown polarity.

\code{bin} aggregates intensity values of chromatograms in discrete bins
along the retention time axis. By default, individual \code{Chromatogram}
objects of one row are binned into the same bins. The function returns a
\code{MChromatograms} object with binned chromatograms.

\code{clean}: removes 0-intensity data points. Either all of them
(with \code{all = TRUE}) or all except those adjacent to non-zero
intensities (\code{all = FALSE}; default). See \code{\link{clean}}
documentation for more details and examples.
}
\details{
The \code{MChromatograms} class extends the base \code{matrix} class
    and hence allows to store \code{\link{Chromatogram}} objects in a
    two-dimensional array. Each row is supposed to contain
    \code{Chromatogram} objects for one MS data \emph{slice} with a common
    m/z and rt range. Columns contain \code{Chromatogram} objects from the
    same sample.

\code{plot}: if \code{nrow(x) > 1} the plot area is split into
    \code{nrow(x)} sub-plots and the chromatograms of one row are plotted in
    each.
}
\note{
Subsetting with \code{[} will always return a \code{MChromatograms}
    object (with the exception of extracting a single element)
    unless \code{drop = TRUE} is specified. This is different from the
    default subsetting behaviour of \code{matrix}-like objects.
}
\examples{
## Creating some chromatogram objects to put them into a MChromatograms object
ints <- abs(rnorm(25, sd = 200))
ch1 <- Chromatogram(rtime = 1:length(ints), ints)
ints <- abs(rnorm(32, sd = 90))
ch2 <- Chromatogram(rtime = 1:length(ints), ints)
ints <- abs(rnorm(19, sd = 120))
ch3 <- Chromatogram(rtime = 1:length(ints), ints)
ints <- abs(rnorm(21, sd = 40))
ch4 <- Chromatogram(rtime = 1:length(ints), ints)

## Create a MChromatograms object with 2 rows and 2 columns
chrs <- MChromatograms(list(ch1, ch2, ch3, ch4), nrow = 2)
chrs

## Extract the first element from the second column. Extracting a single
## element always returns a Chromatogram object.
chrs[1, 2]

## Extract the second row. Extracting a row or column (i.e. multiple elements
## returns by default a list of Chromatogram objects.
chrs[2, ]

## Extract the second row with drop = FALSE, i.e. return a MChromatograms
## object.
chrs[2, , drop = FALSE]

## Replace the first element.
chrs[1, 1] <- ch3
chrs

## Add a pheno data.
pd <- data.frame(name = c("first sample", "second sample"),
    idx = 1:2)
pData(chrs) <- pd

## Column names correspond to the row names of the pheno data
chrs

## Access a column within the pheno data
chrs$name

## Access the m/z ratio for each row; this will be NA for the present
## object
mz(chrs)

## Create some random Chromatogram objects
ints <- abs(rnorm(123, mean = 200, sd = 32))
ch1 <- Chromatogram(rtime = seq_along(ints), intensity = ints, mz = 231)
ints <- abs(rnorm(122, mean = 250, sd = 43))
ch2 <- Chromatogram(rtime = seq_along(ints), intensity = ints, mz = 231)
ints <- abs(rnorm(125, mean = 590, sd = 120))
ch3 <- Chromatogram(rtime = seq_along(ints), intensity = ints, mz = 542)
ints <- abs(rnorm(124, mean = 1200, sd = 509))
ch4 <- Chromatogram(rtime = seq_along(ints), intensity = ints, mz = 542)

## Combine into a 2x2 MChromatograms object
chrs <- MChromatograms(list(ch1, ch2, ch3, ch4), byrow = TRUE, ncol = 2)

## Plot the second row
plot(chrs[2, , drop = FALSE])

## Plot all chromatograms
plot(chrs, col = c("#ff000080", "#00ff0080"))
}
\seealso{
\code{\link{Chromatogram}} for the class representing chromatogram
    data.
    \code{\link{chromatogram}} for the method to extract a
    \code{MChromatograms} object from a \code{\linkS4class{MSnExp}} or
    \code{\linkS4class{OnDiskMSnExp}} object.
    \code{\link{readSRMData}} for the function to read chromatographic data
    of an SRM/MRM experiment.
}
\author{
Johannes Rainer
}
