% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MSnbaseOptions}
\alias{MSnbaseOptions}
\alias{isMSnbaseVerbose}
\title{MSnbase options}
\usage{
MSnbaseOptions()

isMSnbaseVerbose()
}
\value{
A \code{list} of MSnbase options and the single option
    values for the individual accessors.
}

