% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnExperiment-class.R
\docType{class}
\name{MSnExperiment-class}
\alias{MSnExperiment-class}
\alias{readMSnExperiment}
\title{The MSnExperiment class}
\usage{
readMSnExperiment(file, sampleData, backend = BackendMzR(),
  smoothed = NA, BPPARAM = bpparam())
}
\arguments{
\item{file}{Path to mass spectrometry data files}

\item{sampleData}{A [S4Vectors::DataFrame-class] object with additional
information to each sample (samples as rows, information as columns).}

\item{backend}{A [Backend-class] derivate used for internal data storage.}

\item{smoothed}{`logical`, are the spectra smoothed?}

\item{BPPARAM}{Should parallel processing be used? See
[BiocParallel::bpparam()].}
}
\description{
The [MSnExperiment-class] encapsulates data and meta-data for mass
spectrometry experiments.

Draft import method for the new [MSnExperiment-class]. Used as constructor.
}
\details{
In contrast to the old [MSnExp-class] this class supports multiple data
backends, e.g. in-memory ([BackendMemory-class]), on-disk as
mzML ([BackendMzMl-class]) or HDF5 ([BackendHdf5-class]). It supersedes
[MSnExp-class] and [OnDiskMSnExp-class].
}
\section{Slots}{

\describe{
\item{\code{backend}}{A derivate of [Backend-class] holding/controlling the spectra
data.}

\item{\code{spectraData}}{A [S4Vectors::DataFrame-class] storing spectra metadata.}

\item{\code{sampleData}}{A [S4Vectors::DataFrame-class] storing sample metadata.}

\item{\code{processinQueue}}{A `list` storing [ProcessingSteps-class] objects for
lazy processing.}

\item{\code{processing}}{A `character` storing logging information.}

\item{\code{files}}{A `character` storing absolute path to source (in general .mzML)
files.}
}}

\examples{
## TODO
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
