% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnExperiment-class.R
\docType{methods}
\name{MSnExperiment}
\alias{MSnExperiment}
\alias{readMSnExperiment}
\alias{spectrapply,MSnExperiment-method}
\alias{spectra,MSnExperiment-method}
\title{The MSnExperiment class to manage and access MS data}
\usage{
readMSnExperiment(file, sampleData, backend = BackendMzR(),
  smoothed = NA, BPPARAM = bpparam())

\S4method{spectrapply}{MSnExperiment}(object, FUN = NULL,
  BPPARAM = bpparam(), ...)

\S4method{spectra}{MSnExperiment}(object, BPPARAM = bpparam())
}
\arguments{
\item{file}{\code{character} with the file names of the experiment.}

\item{sampleData}{A \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} object with additional
information on each sample (samples as rows, information as columns).}

\item{backend}{A \linkS4class{Backend} derivate used for internal data storage.}

\item{smoothed}{\code{logical}, are the spectra smoothed?}

\item{BPPARAM}{Should parallel processing be used? See
\code{\link[BiocParallel:bpparam]{BiocParallel::bpparam()}}.}
}
\value{
See individual method description for the return value.
}
\description{
The \code{MSnExperiment} class encapsules data and meta-data for mass
spectrometry experiments.

In contrast to the old \linkS4class{MSnExp} this class supports multiple data
backends, e.g. in-memory (\linkS4class{BackendMemory}), on-disk as
mzML (\linkS4class{BackendMzMl}) or HDF5 (\linkS4class{BackendHdf5}). It supersedes
\linkS4class{MSnExp} and \linkS4class{OnDiskMSnExp} objects.
}
\section{Creation of objects}{


\code{MSnExperiment} classes are usually created with the \code{readMSnExperiment}
function that reads general spectrum metadata information from the  mass
spectrometry data files.
}

\section{Accessing data}{

\itemize{
\item \code{spectra}: get the \code{list} of \linkS4class{Spectrum} objects from the experiment.
Note that the spectra in the \code{list} are not grouped by sample/file. Use
the \code{fromFile} method to split/group the \code{list} by file.
\item \code{spectrapply}: apply an arbitrary function to each spectrum in the dataset
and return its result. The function returns a \code{list} with the same length
than there are spectra.
}
}

\section{Subsetting and filtering}{

}

\section{Data manipulation methods}{

}

\examples{

## TODO
}
\author{
Sebastian Gibb, Johannes Rainer
}
