% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-data.R
\name{naplot}
\alias{naplot}
\title{Overview of missing value}
\usage{
naplot(object, verbose = isMSnbaseVerbose(), ...)
}
\arguments{
\item{object}{An object of class \code{MSnSet}.}

\item{verbose}{If verbose (default is \code{TRUE}), print a table
of missing values.}

\item{...}{Additional parameters passed to \code{image2}.}
}
\value{
Used for its side effect. Invisibly returns \code{NULL}
}
\description{
Visualise missing values as a heatmap and barplots along the
samples and features.
}
\examples{
data(naset)
naplot(naset)
}
\author{
Laurent Gatto
}

