\name{numberOfCommonPeaks}
\alias{numberOfCommonPeaks}
\title{number of common peaks}
\usage{
numberOfCommonPeaks(x, y, tolerance = 2.5e-05, relative = TRUE)
}
\arguments{
  \item{x}{spectrum1 (MSnbase::Spectrum)}

  \item{y}{spectrum2 (MSnbase::Spectrum)}

  \item{tolerance}{double, allowed deviation}

  \item{relative}{relative (or absolute) deviation}
}
\value{
double, number of common peaks
}
\description{
number of common peaks
}

