\name{plotSpectrumVsSpectrum}
\alias{plotSpectrumVsSpectrum}
\title{plot spectrum1 vs spectrum2}
\usage{
plotSpectrumVsSpectrum(spectra, tolerance = 0.1, relative = FALSE, ...)
}
\arguments{
  \item{spectra}{list, 2 MSnbase::Spectrum2 objects}

  \item{tolerance}{double, allowed deviation to be
  considered as equal peaks}

  \item{relative}{relative (or absolute) deviation}

  \item{...}{additional paramters passed to
  \code{.plotSpectrumVsSpectrum}.}
}
\description{
plot spectrum1 vs spectrum2
}

