% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\docType{class}
\name{QuantitationParam-class}
\alias{QuantitationParam-class}
\alias{.QuantitationParam}
\alias{QuantitationParam}
\alias{IsobaricTagging}
\alias{SpectralCounting}
\alias{show,QuantitationParam-method}
\alias{IsobaricTagging}
\alias{SpectralCounting}
\title{Quantitation parameter class}
\usage{
\S4method{show}{QuantitationParam}(object)

IsobaricTagging(reporters, msLevel, method = c("max", "trapezoidation",
  "sum"))

SpectralCounting(method = c("count", "SI", "SIgi", "SIn", "SAF", "NSAF"))
}
\arguments{
\item{object}{Object of class `QuantiationParam`.}

\item{reporters}{See class slot.}

\item{msLevel}{See class slot.}

\item{method}{See class slot.}
}
\description{
\code{QuantitationParam} object are used to define the
type of quantitation to perform. These objects are then passed
with the raw data object to the \code{\link[=quantify]{quantify()}}
method. \code{QuantitationParam} objects are rarely created
directly; instead, users should use the predefined constructor
functions \code{IsobaricTagging} and \code{SpectralCounting}.
}
\section{Slots}{

\describe{
\item{\code{msLevel}}{\code{integer(1)} setting the MS level at which
quantitation should be done. It is generally set automatically
for a method (for instance 2 for spectral counting) or from
the data, if absent in the \code{QuantitationParam} object (the
highest MS level in the data, i.e. 2 or 3, for isobaric
tagging).}

\item{\code{reporters}}{The reporter ions to use, as an
\code{\link[=ReporterIons]{ReporterIons()}}. For isobaric tagging quantation only.}

\item{\code{method}}{\code{character(1)} defining the method of
quantitation. For isobaric tagging, this would be one of
\code{"max"} (default), \code{"trapezoidation"}, or \code{"sum"}. For
spectral counting, one of \code{"count"} (default), \code{"SI"},
\code{"SIgi"}, \code{"SIn"}, \code{"SAF"} or \code{"NSAF"}.}

\item{\code{methargs}}{A \code{list} of additional argument applied to the
quantitation method.}

\item{\code{name}}{\code{characater(1)} naming the quantiation parameter.}

\item{\code{wd}}{\code{numeric(1)} defining the width around the experter
isobaric tag mass to seach for a peak. If omitted, is
extracted from the \code{ReporterIons} object. For isobaric tagging
quantation only.}

\item{\code{strict}}{\code{logical(1)} defining if a peak should be quantified
beyond the reporter tag +/- the width. For isobaric tagging
quantation only.}

\item{\code{.__classVersion__}}{The version of the \code{QuantitationParam}
class definition.}
}}

\examples{
## Isobaric tagging using iTRAQ 4-plex
IsobaricTagging(iTRAQ4)
## Isobaric tagging using TMT 11-plex at the MS3 level
IsobaricTagging(TMT11, msLevel = 3L)
## Isobaric tagging using TMT 11-plex at the MS2 level
IsobaricTagging(TMT11, msLevel = 2L)

## Spectral counting, raw counts
SpectralCounting()
## Spectral counting, normalised spectral abundance factor
SpectralCounting("NSAF")
}
\seealso{
quantify
}
\author{
Laurent Gatto
}
