% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{reduce,data.frame-method}
\alias{reduce,data.frame-method}
\title{Reduce a data.frame}
\usage{
\S4method{reduce}{data.frame}(x, key, sep = ";")
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{key}{The column name (currenly only one is supported) to be
used as primary key.}

\item{sep}{The separator. Default is \code{;}.}
}
\value{
A reduced \code{data.frame}.
}
\description{
Reduce a data.frame so that the (primary) key column contains only
unique entries and other columns pertaining to that entry are
combined into semicolon-separated values into a single
row/observation.
}
\examples{
dfr <- data.frame(A = c(1, 1, 2),
                  B = c("x", "y", "z"),
                  C = LETTERS[1:3])
dfr
reduce(dfr, key = "A")
}
\author{
Laurent Gatto
}
