% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumList.R
\name{SpectrumList}
\alias{SpectrumList}
\alias{SpectrumList-class}
\alias{show,SpectrumList-method}
\title{List of Spectrum objects along with annotations}
\usage{
SpectrumList(..., elementMetadata = NULL)
}
\arguments{
\item{...}{\link{Spectrum2} object(s) or a \code{list} of \link{Spectrum2} objects.}

\item{elementMetadata}{\link{DataFrame} with optional information that should
be added as metadata information (\code{mcols}) to the object. The number
of rows has to match the number of \link{Spectrum} objects, each row is
expected to represent additional metadata information for one spectrum.}
}
\description{
\code{SpectrumList} objects allow to collect one or more \link{Spectrum} object(s)
(\link{Spectrum1} or \link{Spectrum2}) in a \code{list}-like structure with the additional
possibility to add arbitrary annotations to each individual \link{Spectrum}
object. These can be accessed/set with the \link{mcols} method.
}
\details{
\code{SpectrumList} inherits all methods from the \link{SimpleList} class of the
\code{S4Vectors} package. This includes \code{lapply} and other data manipulation
and subsetting operations.
}
\section{Constructor}{


New \link{SpectrumList} can be created with the \code{SpectrumList(...)} function
where \code{...} can either be a single \link{Spectrum} object or a \code{list} of
\link{Spectrum} objects (\link{Spectrum1} and/or \link{Spectrum2}).
}

\examples{

## Create from Spectrum objects
sp1 <- new("Spectrum1", mz = c(1, 2, 4), intensity = c(4, 5, 2))
sp2 <- new("Spectrum2", mz = c(1, 2, 3, 4), intensity = c(5, 3, 2, 5),
    precursorMz = 2)

spl <- SpectrumList(sp1, sp2)
spl
spl[[1]]

## Add also metadata columns
mcols(spl)$id <- c("a", "b")
mcols(spl)

## Create a SpectrumList with metadata
spl <- SpectrumList(sp1, sp2, elementMetadata = DataFrame(id = c("a", "b")))

mcols(spl)
mcols(spl)$id
}
\author{
Johannes Rainer
}
