% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitation-MS2-isobaric.R
\name{transferQuantToPrecursorScanNum}
\alias{transferQuantToPrecursorScanNum}
\title{Transfers the quantitation data to it's precusor scan}
\usage{
transferQuantToPrecursorScanNum(x)
}
\arguments{
\item{x}{An \code{\link[=MSnSet]{MSnSet()}}}
}
\value{
An updated \code{\link[=MSnSet]{MSnSet()}}.
}
\description{
This function is meant to be used on isobaric tagging data where
the quantitation and identification are performed on different
scans/MS levels, tpyically MS3 and MS2 respectively.
}
\details{
The output of calling \code{quantify()} on the raw data returns
quantitation feature for all MS3 scans. Calling this function will
move the quantitation data from the MS3 features to the
corresponding MS2 precusor features.

Here's an example illustrating what the function does:
\itemize{
\item MS2 scan with acquisition number 123 is identified as FSEGTSADR,
and what further fragement to quantitation in MS3
\item MS3 scan 125, which originates from scan 123 (i.e. his
\code{precursorScanNum} is 123) has been quantified
\item Then using this function will transfer the quantitation data for
scan 125 to scan 123.
\item As a result, all the quantitation data at the MS3 level in the
input data \code{x} will transfer to MS2 level.
}
}
\seealso{
The \code{\link[=quantify]{quantify()}} methods to generate an \code{\link[=MSnSet]{MSnSet()}}.
}
\author{
Laurent Gatto
}
