\name{updateObject-methods}
\docType{methods}
\alias{updateObject,MSnExp-method}
\alias{updateObject,Spectrum-method}

\title{ Update MSnbase objects }

\description{

  Methods for function \code{updateObject} for objects from the MSnbase
  package.  See \code{\link[BiocGenerics]{updateObject}} for details.
  
}

\section{Methods}{
  \describe{

    \item{\code{signature(object = "MSnExp")}}{ Update the \code{MSnExp}
      object to the latest class version }

    \item{\code{signature(object = "Spectrum")}}{ Update the
      \code{Spectrum} object (and it's sub-classes \code{Spectrum1} and
      \code{Spectrum2}) to the latest class version. }

  }
}

\keyword{methods}

