% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-write.R
\docType{methods}
\name{write,ANY-method}
\alias{write,ANY-method}
\alias{write,MSnExp-method}
\alias{write}
\alias{write,OnDiskMSnExp-method}
\title{Write MS data to mzML or mzXML files}
\usage{
\S4method{write}{ANY}(x, ...)

\S4method{write}{MSnExp}(x, files, outformat = c("mzml", "mzxml"),
  verbose = isMSnbaseVerbose(), copy = TRUE, software_processing = NULL)

\S4method{write}{OnDiskMSnExp}(x, files, outformat = c("mzml", "mzxml"),
  verbose = isMSnbaseVerbose(), copy = TRUE, software_processing = NULL)
}
\arguments{
\item{x}{\code{OnDiskMSnExp} or \code{MSnExp} object.}

\item{...}{For \code{write,ANY}: additional arguments for the \code{\link[base:write]{base::write()}}
function.}

\item{files}{\code{character} with the file name(s). Its length has to match the
number of samples/files of \code{x}.}

\item{outformat}{\code{character(1)} defining the format of the output files.
Default output format is \code{"mzml"}.}

\item{verbose}{\code{logical(1)} if progress messages should be displayed.}

\item{copy}{\code{logical(1)} if metadata (data processings, original file names
etc) should be copied from the original files. See details for more
information.}

\item{software_processing}{optionally provide specific data processing steps.
See documentation of the \code{software_processing} parameter of
\code{\link[mzR:writeMSData]{mzR::writeMSData()}}.}
}
\description{
The \code{write,MSnExp} and \code{write,OnDiskMSnExp} saves the content
of a \link{MSnExp} or \link{OnDiskMSnExp} object to MS file(s) in either
\emph{mzML} or \emph{mzXML} format (one file per sample).
}
\details{
The \code{write} method uses the \emph{proteowizard} libraries through the
\code{mzR} package to save the MS data. The data can be written to \emph{mzML} or
\emph{mzXML} files with or without copying additional metadata information
from the original files from which the data was read by the
\code{\link[=readMSData]{readMSData()}} function. This can be set using the \code{copy} parameter.
Note that \code{copy = TRUE} requires the original files to be available.\preformatted{The `write,ANY` method aims to restore the use of the [base::write()]
function if `x` is neither a `MSnExp` nor an `OnDiskMSnExp` object.
}
}
\note{
General spectrum data such as total ion current, peak count, base peak
m/z or base peak intensity are calculated from the actual spectrum data
before writing the data to the files.\preformatted{For MSn data, if the `OnDiskMSnExp` or `MSnExp` does not contain also the
precursor scan of a MS level > 1 spectrum `precursorMZ`,
`precursorCharge` and `precursorIntensity` of the spectrum is set to 0
in the output file.
}
}
\author{
Johannes Rainer
}
