\name{xic-methods}
\docType{methods}
\alias{xic-methods}
\alias{xic,mzRramp-method}
\alias{xic,character-method}
\alias{xic}

\title{ Methods for Function \code{xic} }

\description{
}

\section{Methods}{
  \describe{
    \item{\code{signature(object = "mzRramp")}}{ }
    \item{\code{signature(object = "character")}}{ }
  }
}
\keyword{methods}

