% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend-class.R
\name{backendWriteSpectra}
\alias{backendWriteSpectra}
\title{Write spectra data to a backend}
\usage{
backendWriteSpectra(object, files, ids, spectra)
}
\arguments{
\item{object}{An object inheriting from [Backend-class],
i.e. [BackendHdf5-class]}

\item{files}{The path to the source (generally .mzML) files.}

\item{ids}{The spectrum indicies.}

\item{spectra}{A list of [Spectrum-class] objects that should be written to
the backend.}
}
\value{
A [Backend-class] derivate.
}
\description{
This generic is used to write spectra data to a backend.
}
\details{
It *MUST* be reimplemented by all backends!
}
\seealso{
Other Backend generics: \code{\link{backendImportData}},
  \code{\link{backendInitialize}},
  \code{\link{backendReadSpectra}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\concept{Backend generics}
