% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MSnExp.R
\docType{methods}
\name{chromatogram,MSnExp-method}
\alias{chromatogram,MSnExp-method}
\alias{chromatogram}
\title{Extract chromatogram object(s)}
\usage{
\S4method{chromatogram}{MSnExp}(object, rt, mz, aggregationFun = "sum",
  missing = NA_real_)
}
\arguments{
\item{object}{For \code{chromatogram}: a \code{\linkS4class{MSnExp}} or
\code{\linkS4class{OnDiskMSnExp}} object from which the chromatogram
should be extracted.}

\item{rt}{A \code{numeric(2)} or two-column \code{matrix} defining the lower
and upper boundary for the retention time range/window(s) for the
chromatogram(s). If a \code{matrix} is provided, a chromatogram is
extracted for each row. If not specified, a chromatogram representing the
full retention time range is extracted. See examples below for details.}

\item{mz}{A \code{numeric(2)} or two-column \code{matrix} defining the
mass-to-charge (mz) range(s) for the chromatogram(s). For each
spectrum/retention time, all intensity values within this mz range are
aggregated to result in the intensity value for the spectrum/retention
time. If not specified, the full mz range is considered. See examples
below for details.}

\item{aggregationFun}{\code{character} defining the function to be used for
intensity value aggregation along the mz dimension. Allowed values are
\code{"sum"} (TIC), \code{"max"} (BPC), \code{"min"} and \code{"mean"}.}

\item{missing}{\code{numeric(1)} allowing to specify the intensity value for
if for a given retention time (spectrum) no signal was measured within
the mz range. Defaults to \code{NA_real_}.}
}
\value{
\code{chromatogram} returns a \code{\link{Chromatograms}} object with
    the number of columns corresponding to the number of files in
    \code{object} and number of rows the number of specified ranges (i.e.
    number of rows of matrices provided with arguments \code{mz} and/or
    \code{rt}).
}
\description{
The \code{chromatogram} method extracts chromatogram(s) from an
    \code{\linkS4class{MSnExp}} or \code{\linkS4class{OnDiskMSnExp}} object.
    Depending on the provided parameters this can be a total ion chromatogram
    (TIC), a base peak chromatogram (BPC) or an extracted ion chromatogram
    (XIC) extracted from each sample/file.
}
\details{
Arguments \code{rt} and \code{mz} allow to specify the MS
    data slice from which the chromatogram should be extracted.
    The parameter \code{aggregationSum} allows to specify the function to be
    used to aggregate the intensities across the mz range for the same
    retention time. Setting \code{aggregationFun = "sum"} would e.g. allow
    to calculate the \emph{total ion chromatogram} (TIC),
    \code{aggregationFun = "max"} the \emph{base peak chromatogram} (BPC).
    The length of the extracted \code{\link[MSnbase]{Chromatogram}} object,
    i.e. the number of available data points, corresponds to the number of
    scans/spectra measured in the specified retention time range. If in a
    specific scan (for a give retention time) no signal was measured in the
    specified mz range, a \code{NA_real_} is reported as intensity for the
    retention time (see Notes for more information). This can be changed
    using the \code{missing} parameter.

    By default or if \code{mz} and/or \code{rt} are numeric vectors, the
    function extracts one \code{\link{Chromatogram}} object for each file
    in the \code{\linkS4class{MSnExp}} or \code{\linkS4class{OnDiskMSnExp}}
    object. Providing a numeric matrix with argument \code{mz} or \code{rt}
    enables to extract multiple chromatograms per file, one for each row in
    the matrix. If the number of columns of \code{mz} or \code{rt} are not
    equal to 2, \code{range} is called on each row of the matrix.
}
\author{
Johannes Rainer
}
