% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-MSnExp.R
\name{combineSpectraMovingWindow}
\alias{combineSpectraMovingWindow}
\title{Combine signal from consecutive spectra of LCMS experiments}
\usage{
combineSpectraMovingWindow(x, halfWindowSize = 1L, mzFun = base::mean,
  intensityFun = base::sum, mzd, BPPARAM = bpparam())
}
\arguments{
\item{x}{\code{MSnExp} or \code{OnDiskMSnExp} object.}

\item{halfWindowSize}{\code{integer(1)} with the half window size for the moving
window.}

\item{mzFun}{\code{function} to aggregate the m/z values per m/z group.}

\item{intensityFun}{\code{function} to aggregate the intensity values per m/z
group.}

\item{mzd}{\code{numeric(1)} defining the maximal m/z difference below which
values are grouped. If not provided, this value is estimated from the
distribution of differences of m/z values from the provided spectra
(see details).}

\item{BPPARAM}{parallel processing settings.}
}
\value{
\code{MSnExp} with the same number of spectra than \code{x}.
}
\description{
\code{combineSpectraMovingWindow} combines signal from consecutive spectra within
a file. The resulting \code{MSnExp} has the same total number of spectra than the
original object, but with each individual's spectrum information
representing aggregated data from the original spectrum and its neighboring
spectra.

Note that the function returns always a \code{MSnExp} object, even if \code{x} was an
\code{OnDiskMSnExp} object.
}
\details{
The method assumes same ions being measured in consecutive scans (i.e. LCMS
data) and thus combines their signal which can increase the increase the
signal to noise ratio.

See \code{\link[=combineSpectra]{combineSpectra()}} for details.
}
\seealso{
\code{\link[=combineSpectra]{combineSpectra()}} for the function combining spectra provided in
a \code{list}.

\code{\link[=estimateMzScattering]{estimateMzScattering()}} for a function to estimate m/z value scattering in
consecutive spectra.
}
\author{
Johannes Rainer
}
