\name{Deprecated}
\title{MSnbase Deprecated and Defunct}

\alias{MSnbase-deprecated}
\alias{MSnbase-defunct}

\alias{readMzXMLData}

\description{
  The function, class, or data object you have asked for has been
  deprecated or made defunct. 

  Deprecated: 

  Defunct: \code{readMzXMLData}, \code{extractSpectra}.
}

\keyword{internal}
