\name{Deprecated}
\title{MSnbase Deprecated and Defunct}

\alias{MSnbase-deprecated}
\alias{MSnbase-defunct}

\alias{readMzXMLData}
\alias{extractSpectra}
\alias{writeMzTabData}
\alias{makeMTD}
\alias{makePEP}
\alias{makePRT}

\description{
  The function, class, or data object you have asked for has been
  deprecated or made defunct. 

  Deprecated: 

  Defunct: \code{readMzXMLData}, \code{extractSpectra},
  \code{writeMzTabData}, \code{makeMTD}, \code{makePEP}, \code{makePRT}.
  
}

\keyword{internal}
