\name{Deprecated}
\title{MSnbase Deprecated and Defunct}

\alias{MSnbase-deprecated}
\alias{MSnbase-defunct}

\alias{readMzXMLData}
\alias{extractSpectra}
\alias{makeMTD}
\alias{makePEP}
\alias{makePRT}

\alias{NAnnotatedDataFrame-class}
\alias{NAnnotatedDataFrame}
\alias{class:NAnnotatedDataFrame}
\alias{multiplex,NAnnotatedDataFrame-method}
\alias{multiLabels,NAnnotatedDataFrame-method}
\alias{dim,NAnnotatedDataFrame-method}
\alias{show,NAnnotatedDataFrame-method}
\alias{multiplex}
\alias{multiLabels}

\description{
  The function, class, or data object you have asked for has been
  deprecated or made defunct. 

  Deprecated: 

  Defunct: \code{readMzXMLData}, \code{extractSpectra},
  \code{writeMzTabData}, \code{makeMTD}, \code{makePEP}, \code{makePRT},
  \code{NAnnotatedDataFrame} class.
  
}

\keyword{internal}
