% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatRt}
\alias{formatRt}
\title{Format Retention Time}
\usage{
formatRt(rt)
}
\arguments{
\item{rt}{retention time in seconds (\code{numeric}) or "mm:sec"
(\code{character}).}
}
\value{
A vector of same length as \code{rt}.
}
\description{
This function is used to convert retention times. Conversion is
seconds to/from the more human friendly format "mm:sec". The
implementation is from \code{\link[MsCoreUtils:formatRt]{MsCoreUtils::formatRt()}}.
}
\examples{

formatRt(1524)
formatRt("25:24")
}
\author{
Laurent Gatto and Sebastian Gibb
}
