% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.atomic.mass}
\alias{get.atomic.mass}
\title{Atomic mass.}
\usage{
get.atomic.mass()
}
\value{
A named \code{double}.
}
\description{
Returns a \code{double} of used atomic mass.
}
\examples{
get.atomic.mass()
}
\author{
Sebastian Gibb
}
