% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\docType{class}
\name{Hdf5MSnExp-class}
\alias{Hdf5MSnExp-class}
\alias{.Hdf5MSnExp}
\alias{Hdf5MSnExp}
\alias{readHdf5DiskMSData}
\alias{hdf5Close}
\alias{hdf5Open}
\alias{isHdf5Open}
\title{The \code{Hdf5MSnExp} Class for MS Data And Meta-Data}
\usage{
readHdf5DiskMSData(files, pdata = NULL, msLevel. = NULL,
  verbose = isMSnbaseVerbose(), centroided. = NA, smoothed. = NA,
  hdf5file = NULL, openHdf5 = TRUE)

hdf5Close(object)

hdf5Open(object)

isHdf5Open(object)
}
\arguments{
\item{object}{An instance of class `Hdf5MSnExp`.}
}
\description{
The \code{Hdf5MSnExp} class encapsulates data and
meta-data for mass spectrometry experiments like the \code{MSnExp}
and \code{OnDiskMSnExp} classes. \code{Hdf5MSnExp} implements, like
\code{OnDiskMSnExp}, an \emph{on disk} model, where the raw data (M/Z
and intensities) are stored on disk (rather than in memory,
like \code{MSnExp} data) using, as the name implies, the hdf5 file
format. See \code{\link{MSnExp}} and \code{\link{OnDiskMSnExp}} for a more general
description of the data, slots and operations.\preformatted{Object of the class are currently created with the
`readHdf5DiskMSData` function. Later, this backend will be
included in the main `readMSData` function.
}
}
\section{Slots}{

\describe{
\item{\code{hdf5file}}{\code{character(1)} containing the hdf5 filename.}

\item{\code{hdf5handle}}{A \code{H5IdComponent} or \code{NULL}.}
}}

\examples{
f <- msdata::proteomics(pattern = "MS3TMT11", full.names = TRUE)
x <- readHdf5DiskMSData(f)
x
x[[1]]
x[[2]]
x[[10]]
filterMsLevel(x, 3L)[[1]]
}
\seealso{
See \code{\link{MSnExp}} and \code{\link{OnDiskMSnExp}} classes.
}
\author{
Laurent Gatto
}
