% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/Backend-class.R,
%   R/BackendHdf5-class.R, R/BackendMemory-class.R, R/BackendMzR-class.R,
%   R/MSnExperiment-class.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{show,Backend-method}
\alias{fileNames,Backend-method}
\alias{backendInitialize,Backend-method}
\alias{backendImportData,Backend-method}
\alias{backendDeepCopy,Backend-method}
\alias{backendSubset}
\alias{show,BackendHdf5-method}
\alias{backendSubset,BackendMemory-method}
\alias{backendInitialize,BackendMemory-method}
\alias{backendImportData,BackendMemory-method}
\alias{backendReadSpectra,BackendMemory-method}
\alias{backendWriteSpectra,BackendMemory-method}
\alias{backendReadSpectra,BackendMzR-method}
\alias{show,MSnExperiment-method}
\alias{setBackend,MSnExperiment,Backend-method}
\alias{setBackend,MSnExperiment,BackendMzR-method}
\alias{setBackend,MSnExperiment,BackendHdf5-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{Backend}(object)

\S4method{fileNames}{Backend}(object, ...)

\S4method{backendInitialize}{Backend}(object, files, spectraData, ...)

\S4method{backendImportData}{Backend}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendDeepCopy}{Backend}(object, ...)

backendSubset(object, spectraData)

\S4method{show}{BackendHdf5}(object)

\S4method{backendSubset}{BackendMemory}(object, spectraData)

\S4method{backendInitialize}{BackendMemory}(object, files, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendImportData}{BackendMemory}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendReadSpectra}{BackendMemory}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendWriteSpectra}{BackendMemory}(object, spectra, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendReadSpectra}{BackendMzR}(object, spectraData,
  processingQueue = list(), ...)

\S4method{show}{MSnExperiment}(object)

\S4method{setBackend}{MSnExperiment,Backend}(object, backend, ...,
  BPPARAM = bpparam())

\S4method{setBackend}{MSnExperiment,BackendMzR}(object, backend, ...,
  BPPARAM = bpparam())

\S4method{setBackend}{MSnExperiment,BackendHdf5}(object, backend, ...,
  BPPARAM = bpparam())
}
\arguments{
\item{object}{Object to display.}

\item{spectraData}{`DataFrame` with the spectrum metadata of the spectra to
which the `object` should be subsetted.}

\item{x}{`Backend`}

\item{object}{Object to display.}
}
\value{
A `Backend` class.
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\author{
Johannes Rainer
}
\keyword{internal}
