% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/Backend-class.R,
%   R/BackendMemory-class.R, R/MSnExperiment-class.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{show,Backend-method}
\alias{backendInitialize,Backend-method}
\alias{backendImportData,Backend-method}
\alias{backendDeepCopy,Backend-method}
\alias{backendImportData,BackendMemory-method}
\alias{backendReadSpectra,BackendMemory-method}
\alias{backendWriteSpectra,BackendMemory-method}
\alias{show,MSnExperiment-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{Backend}(object)

\S4method{backendInitialize}{Backend}(object, files, ...)

\S4method{backendImportData}{Backend}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendDeepCopy}{Backend}(object, ..., BPPARAM = bpparam())

\S4method{backendInitialize}{Backend}(object, files, ...)

\S4method{backendImportData}{BackendMemory}(object, files, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendReadSpectra}{BackendMemory}(object, file, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendWriteSpectra}{BackendMemory}(object, file, spectra,
  spectraData, ..., BPPARAM = bpparam())

\S4method{show}{MSnExperiment}(object)
}
\arguments{
\item{object}{Object to display.}

\item{object}{Object to display.}
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
