% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/Backend-class.R,
%   R/BackendMemory-class.R, R/BackendMzR-class.R, R/MSnExperiment-class.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{show,Backend-method}
\alias{fileNames,Backend-method}
\alias{filterFile,Backend-method}
\alias{backendInitialize,Backend-method}
\alias{backendImportData,Backend-method}
\alias{backendDeepCopy,Backend-method}
\alias{[,BackendMemory,ANY,missing,ANY-method}
\alias{filterFile,BackendMemory-method}
\alias{backendInitialize,BackendMemory-method}
\alias{backendImportData,BackendMemory-method}
\alias{backendReadSpectra,BackendMemory-method}
\alias{backendWriteSpectra,BackendMemory-method}
\alias{backendReadSpectra,BackendMzR-method}
\alias{show,MSnExperiment-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{Backend}(object)

\S4method{fileNames}{Backend}(object, ...)

\S4method{filterFile}{Backend}(object, file, ...)

\S4method{backendInitialize}{Backend}(object, files, spectraData, ...)

\S4method{backendImportData}{Backend}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendDeepCopy}{Backend}(object, ...)

\S4method{[}{BackendMemory,ANY,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{filterFile}{BackendMemory}(object, file, ...)

\S4method{backendInitialize}{BackendMemory}(object, files, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendImportData}{BackendMemory}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendReadSpectra}{BackendMemory}(object, spectraData, ...,
  BPPARAM = bpparam())

\S4method{backendWriteSpectra}{BackendMemory}(object, spectra, spectraData,
  ..., BPPARAM = bpparam())

\S4method{backendReadSpectra}{BackendMzR}(object, spectraData,
  processingQueue = list(), ...)

\S4method{show}{MSnExperiment}(object)
}
\arguments{
\item{object}{Object to display.}

\item{object}{Object to display.}
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
