% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iPQF.R
\name{iPQF}
\alias{iPQF}
\title{iPQF: iTRAQ (and TMT) Protein Quantification based on Features}
\usage{
iPQF(object, groupBy, low.support.filter = FALSE, ratio.calc = "sum",
  method.combine = FALSE)
}
\arguments{
\item{object}{An instance of class \code{MSnSet} containing
absolute ion intensities.}

\item{groupBy}{Vector defining spectra to protein
matching. Generally, this is a feature variable such as
\code{fData(object)$accession}.}

\item{low.support.filter}{A \code{logical} specifying if proteins
being supported by only 1-2 peptides should be filtered
out. Default is \code{FALSE}.}

\item{ratio.calc}{Either \code{"none"} (don't calculate any
ratios), \code{"sum"} (default), or a specific channel (one of
\code{sampleNames(object)}) defining how to calculate relative
peptides intensities.}

\item{method.combine}{A \code{logical} defining whether to further
use median polish to combine features.}
}
\value{
A \code{matrix} with estimated protein ratios.
}
\description{
The iPQF spectra-to-protein summarisation method integrates
peptide characteristics and quantitative values for protein
quantitation estimation, were peptide spectra receive weights and
contribute to the protein quantification according to their
reliability. See also \code{\link{combineFeatures}} for a more
general overview of feature aggregation and examples.
}
\author{
Martina Fisher
}
\references{
iPQF: A new peptide-to-protein summarization method
using peptide characteristics to improve iTRAQ quantification
Martina Fischer and Bernhard Y. Renard, \emph{in prep}.
}

