% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listOf}
\alias{listOf}
\title{Tests equality of list elements class}
\usage{
listOf(x, class, valid = TRUE)
}
\arguments{
\item{x}{A \code{list}.}

\item{class}{A \code{character} defining the expected class.}

\item{valid}{A \code{logical} defining if all elements should be
tested for validity. Default is \code{TRUE}.}
}
\value{
\code{TRUE} is all elements of \code{x} inherit from
\code{class}.
}
\description{
Compares equality of all members of a list.
}
\examples{
listOf(list(), "foo")
listOf(list("a", "b"), "character")
listOf(list("a", 1), "character")
}
\author{
Laurent Gatto
}
