% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{MSnbaseOptions}
\alias{MSnbaseOptions}
\alias{isMSnbaseVerbose}
\alias{setMSnbaseParallelThresh}
\alias{setMSnbaseVerbose}
\title{MSnbase options}
\usage{
MSnbaseOptions()

isMSnbaseVerbose()

setMSnbaseVerbose(opt)

setMSnbaseParallelThresh(opt = 1000)
}
\arguments{
\item{opt}{The value of the new option}
}
\value{
A \code{list} of MSnbase options and the single option
    values for the individual accessors.
}
\description{
MSnbase defined a few options globally using the standard R
options mechanism. The current values of these options can be
queried with \code{MSnbaseOptions}. The options are
\code{verbose}, that defines a session-wide verbosity flag, that
is used if the \code{verbose} argument in individual functions is
not set. The \code{PARALLEL_THRESH} defines the minimum number of
spectra per file necessary before using parallel processing.
}
\details{
\code{isMSnbaseVerbose} is one wrapper for the verbosity flag,
also available through \code{options("MSnbase")$verbose}.

There are also setters to set options individually. When run
without argument, the verbosity setter inverts the current value
of the option.
}

