% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnExperiment-class.R
\docType{class}
\name{MSnExperiment-class}
\alias{MSnExperiment-class}
\title{The MSnExperiment class}
\description{
The [MSnExperiment-class] encapsulates data and meta-data for mass
spectrometry experiments.
}
\details{
In contrast to the old [MSnExp-class] this class supports multiple data
backends, e.g. in-memory ([BackendMemory-class]), on-disk as
mzML ([BackendMzMl-class]) or HDF5 ([BackendHdf5-class]). It supersedes
[MSnExp-class] and [OnDiskMSnExp-class].
}
\section{Slots}{

\describe{
\item{\code{backend}}{A derivate of [Backend-class] holding/controlling the spectra
data.}

\item{\code{featureData}}{A [S4Vectors::DataFrame-class] storing spectra metadata.}

\item{\code{phenoData}}{A [S4Vectors::DataFrame-class] storing sample metadata.}

\item{\code{processinQueue}}{A `list` storing [ProcessingSteps-class] objects for
lazy processing.}

\item{\code{processing}}{A `character` storing logging information.}

\item{\code{files}}{A `character` storing absolute path to source (in general .mzML)
files.}
}}

\examples{
## TODO
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
