% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnExperiment-class.R
\docType{methods}
\name{MSnExperiment}
\alias{MSnExperiment}
\alias{MSnExperiment-class}
\alias{readMSnExperiment}
\alias{spectrapply,MSnExperiment-method}
\alias{spectra,MSnExperiment-method}
\alias{removePeaks,MSnExperiment-method}
\alias{clean,MSnExperiment-method}
\title{The MSnExperiment class to manage and access MS data}
\usage{
readMSnExperiment(file, sampleData, backend = BackendMzR(),
  smoothed = NA, BPPARAM = bpparam())

\S4method{spectrapply}{MSnExperiment}(object, FUN = NULL,
  BPPARAM = bpparam(), ...)

\S4method{spectra}{MSnExperiment}(object, BPPARAM = bpparam())

\S4method{removePeaks}{MSnExperiment}(object, t = "min",
  verbose = isMSnbaseVerbose(), msLevel.)

\S4method{clean}{MSnExperiment}(object, all = FALSE,
  verbose = isMSnbaseVerbose(), msLevel.)
}
\arguments{
\item{file}{\code{character} with the file names of the experiment.}

\item{sampleData}{A \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} object with additional
information on each sample (samples as rows, information as columns).}

\item{backend}{A \linkS4class{Backend} derivate used for internal data storage.}

\item{smoothed}{\code{logical}, are the spectra smoothed?}

\item{BPPARAM}{Should parallel processing be used? See
\code{\link[BiocParallel:bpparam]{BiocParallel::bpparam()}}.}

\item{object}{A \code{MSnExperiment} object.}

\item{FUN}{for \code{spectrapply}: a function or the name of a function to apply
to each \linkS4class{Spectrum} of the experiment.}

\item{...}{for \code{spectrapply}: additional arguments to be passed to \code{FUN}.}

\item{t}{for \code{removePeaks}: a \code{numeric(1)} defining the threshold or \code{"min"}.}

\item{verbose}{\code{logical(1)} defining the verbosity.}

\item{msLevel.}{\code{integer} defining the MS level of the spectra to which the
function should be applied.}

\item{all}{for \code{clean}: \code{logical(1)} whether all 0 intensity peaks should be
removed (\code{TRUE}) or whether 0-intensity peaks directly adjacent to a
non-zero intensity peak should be kept (\code{FALSE}).}
}
\value{
See individual method description for the return value.
}
\description{
The \code{MSnExperiment} class encapsules data and meta-data for mass
spectrometry experiments.

In contrast to the old \linkS4class{MSnExp} this class supports multiple data
backends, e.g. in-memory (\linkS4class{BackendMemory}), on-disk as
mzML (\linkS4class{BackendMzMl}) or HDF5 (\linkS4class{BackendHdf5}). It supersedes
\linkS4class{MSnExp} and \linkS4class{OnDiskMSnExp} objects.
}
\details{
The \code{MSnExperiment} class uses by default a lazy data manipulation strategy,
i.e. data manipulations such as with methods \code{removePeaks} are not applied
immediately to the data, but applied on-the-fly to the spectrum data once it
is retrieved.
}
\section{Creation of objects}{


\code{MSnExperiment} classes are usually created with the \code{readMSnExperiment}
function that reads general spectrum metadata information from the  mass
spectrometry data files.
}

\section{Accessing data}{

\itemize{
\item \code{spectra}: get the \code{list} of \linkS4class{Spectrum} objects from the experiment.
Note that the spectra in the \code{list} are not grouped by sample/file. Use
the \code{fromFile} method to split/group the \code{list} by file.
\item \code{spectrapply}: apply an arbitrary function to each spectrum in the dataset
and return its result. The function returns a \code{list} with the same length
than there are spectra.
}
}

\section{Subsetting and filtering}{

}

\section{Data manipulation methods}{

\itemize{
\item \code{clean}: remove 0-intensity data points. See \code{\link[=clean]{clean()}} for
\linkS4class{Spectrum} objects for more details.
\item \code{removePeaks}: remove peaks lower than a threshold \code{t}. See
\code{\link[=removePeaks]{removePeaks()}} for \linkS4class{Spectrum} objects for more details.
}
}

\examples{

## TODO
}
\author{
Sebastian Gibb, Johannes Rainer
}
