% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitation.R
\name{quantify-methods}
\alias{quantify-methods}
\alias{quantify}
\title{Quantifies raw MS data objects}
\arguments{
\item{object}{Raw mass spectrometry data to be quantified. Can be
a single \code{\link[=Spectrum]{Spectrum()}} or a complete experiment provided as
\code{\link[=MSnExp]{MSnExp()}} or \code{\link[=OnDiskMSnExp]{OnDiskMSnExp()}}.}

\item{method}{An instance of \code{QuantitationParam} defining the
quantitation method. For isobaric tagging, the method
quantifies peaks defined in \code{\link[=IsobaricTagging]{IsobaricTagging()}} object. For
label-free quantitation, defined by a \code{\link[=SpectralCounting]{SpectralCounting()}}
quantitation parameter, the respective quantitation methods
and normalisations are applied to the spectra. These methods
require two additional arguments passed in \code{...}, namely the
protein accession of identifiers (\code{fcol}, with detault value
\code{"DatabaseAccess"}) and the protein lengths (\code{plength}, with
default value \code{"DBseqLength"}). These values are available of
the identification data had been collated using
\code{\link[=addIdentificationData]{addIdentificationData()}}.\preformatted{See [QuantitationParam()] for additional details on how to
define quantiation parameters.
}}

\item{BPPARAM}{Support for parallel processing using the
\code{BiocParallel} infrastructure. When missing (default), the
default registered \code{BiocParallelParam} parameters are applied
using \code{bpparam()}. Alternatively, one can pass a valid
\code{BiocParallelParam} parameter instance: \code{SnowParam},
\code{MulticoreParam}, \code{DoparParam}, ... see the
\code{BiocParallel} package for details.}

\item{pepseq}{\code{character(1)} giving the peptide sequence column in
the feature data. Default is \code{"sequence"}.}

\item{verbose}{\code{logical(1)} defining whether a progress bar should
be displayed when quantifying full experiments. Default is
\code{\link[=isMSnbaseVerbose]{isMSnbaseVerbose()}}.}

\item{...}{Further arguments passed to the quantitation functions.}
}
\value{
An object of class \code{\link[=MSnSet]{MSnSet()}} is returned containing the
quantified feature expression and all meta data inherited from
the mass spectrometry experiment data. When quantifying
individual spectra with isobaric tagging, a list of length 2
will be returned. The first element, named \code{peakQuant}, is a
\code{numeric} of length equal to the number of reportesr with
quantitation of the reporter peaks.\preformatted{The second element, named `curveStats`, is a `data.frame`
providing, for each reporter, its curve parameters: maximum
intensity *maxInt*, number of maxima *nMaxInt*, the number of
data points defined the curve *baseLength*, lower and upper MZ
values for the curve (*lowerMz* and *upperMz*), and reporter
reporter and precursor MZ values (when available).
}
}
\description{
These methods quantify individual spectra or full MS
experiments. Currently, MS2-level isobar tagging using iTRAQ
and TMT (or any arbitrary peaks of interest, see
\code{\link[=ReporterIons]{ReporterIons()}} and MS2-level label-free quantitation
(spectral counting, spectral index or spectral abundance
factor) are available, as described in
\code{\link[=QuantitationParam]{QuantitationParam()}}.
}
\details{
\code{\link[=ReporterIons]{ReporterIons()}} define specific MZ at which peaks are
expected and a window around that MZ value. A peak of interest
is searched for in that window.  Once the range of the curve
is found, quantification is performed.  If no data points are
found in the expected region, \code{NA} is returned for the
reporter peak MZ.\preformatted{For spectral counting, spectra that have not been identified
(the corresponding fields in the feature data are populated
with `NA` values) or that have not been uniquely assigned to a
protein (the `nprot` feature variable is greater that 1) are
removed prior to quantitation. The latter does not apply for
method `"count"` (see [QuantitationParam()]) but can be
applied manually with [removeMultipleAssignment()].
}
}
\examples{
## Quantifying a full experiment using iTRAQ4-plex tagging
data(itraqdata)
msnset <- quantify(itraqdata, method = "trap", reporters = iTRAQ4)
msnset

## specifying a custom parallel framework
## bp <- MulticoreParam(2L) # on Linux/OSX
## bp <- SnowParam(2L) # on Windows
## quantify(itraqdata[1:10], method = "trap", iTRAQ4, BPPARAM = bp)

## Checking for non-quantified peaks
sum(is.na(exprs(msnset)))

## Quantifying a single spectrum
qty <- quantify(itraqdata[[1]], method = "trap", iTRAQ4[1])
qty$peakQuant
qty$curveStats

## Label-free quantitation
## Raw (mzXML) and identification (mzid) files
quantFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "mzXML$")
identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                 full.name = TRUE, pattern = "dummyiTRAQ.mzid")
msexp <- readMSData(quantFile)
msexp <- addIdentificationData(msexp, identFile)
fData(msexp)$DatabaseAccess

si <- quantify(msexp, method = "SIn")
processingData(si)
exprs(si)

saf <- quantify(msexp, method = "NSAF")
processingData(saf)
exprs(saf)
}
\seealso{
The legacy interface to \code{quantify} is still availabel in
\link{quantify-legacy}.
}
\author{
Laurent Gatto and Sebastian Gibb
}
