% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions-MSnSet.R
\name{selectFeatureData}
\alias{selectFeatureData}
\title{Select feature variables of interest}
\usage{
selectFeatureData(object, graphics = TRUE)
}
\arguments{
\item{object}{An \code{MSnSet}.}

\item{graphics}{A \code{logical} (default is \code{TRUE})
indicating whether a shiny application should be used if
available. Otherwise, a text menu is used.}
}
\value{
Updated \code{MSnSet}, containing only selected feature
    variables.
}
\description{
Select feature variables to be retained.
}
\examples{
\dontrun{
library("pRolocdata")
data(hyperLOPIT2015)
x <- selectFeatureData(hyperLOPIT2015) ## select via GUI
head(fData(x))
}
}
\author{
Laurent Gatto
}

