% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-MSnSet.R
\name{selectFeatureData}
\alias{selectFeatureData}
\title{Select feature variables of interest}
\usage{
selectFeatureData(object, graphics = TRUE, fcol)
}
\arguments{
\item{object}{An \code{MSnSet}.}

\item{graphics}{A \code{logical} (default is \code{TRUE})
indicating whether a shiny application should be used if
available. Otherwise, a text menu is used. Ignored if \code{k}
is not missing.}

\item{fcol}{A \code{numeric}, \code{local} or \code{character} of
valid feature variables to be passed directly.}
}
\value{
Updated \code{MSnSet}, containing only selected feature
    variables.
}
\description{
Select feature variables to be retained.
}
\examples{
library("pRolocdata")
data(hyperLOPIT2015)
## 5 first feature variables
x <- selectFeatureData(hyperLOPIT2015, fcol = 1:5)
fvarLabels(x)
\dontrun{
## select via GUI
x <- selectFeatureData(hyperLOPIT2015)
fvarLabels(x)
}
}
\author{
Laurent Gatto
}

