% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readWriteMzTab.R
\name{writeMzTabData}
\alias{writeMzTabData}
\title{Writes an 'MSnSet' to an mzTab file}
\usage{
writeMzTabData(x, what = c("PEP", "PRT"), append = FALSE, MTD = TRUE,
  file, ...)
}
\arguments{
\item{x}{An instance of class \code{MSnSet}.}

\item{what}{One of \code{"PEP"} or \code{"PRT"} defining whether
peptide or protein data is to be saved.}

\item{append}{Logical. Should the data be appended to \code{file}.
Default is \code{FALSE}.}

\item{MTD}{Logical. Should the metadata section be generated.
Default is \code{TRUE}.}

\item{file}{A \code{character} naming the file to print to.}

\item{...}{Additional parameters passed to the respective section
generators: \code{\link{makeMTD}} for metadata, \code{\link{makePEP}}
for peptides and \code{\link{makePRT}} for proteins.}
}
\value{
None (invisible \code{NULL}).
}
\description{
This function generates an \code{mzTab} file based on the data
available in teh \code{x} \code{MSnSet} instance and additional
information passed by the user. It make used of the respective
section generators to create appropriate metadata, peptide and
protein sections. If peptide and protein sections need to be
generated, one has to first create the \code{mzTab} file with
(metadata, optional, but recommended and) protein data and then
append the peptide data, as per \code{mzTab} specification
(\url{http://code.google.com/p/mztab/}). See the example section.
}
\examples{
mzTabFile <- tempfile()
data(msnset)
pep <- msnset
prot <- combineFeatures(pep, groupBy = fData(pep)$ProteinAccession)
fvarLabels(pep)
## First write metadata and protein data
writeMzTabData(prot, what = "PRT", file = mzTabFile,
               append = FALSE, MTD = TRUE,
               protAccession = fData(prot)$ProteinAccession,
               protDescription = fData(prot)$ProteinDescription,
               protAbundance = exprs(prot))
## append peptide data, without metadata section
writeMzTabData(pep, what = "PEP", file = mzTabFile,
               append = TRUE, MTD = FALSE,
               sequence = fData(pep)$PeptideSequence,
               charge = fData(pep)$charge,
               retentionTime = fData(pep)$retention.time,
               pepAbundance = exprs(pep))
}
\author{
Laurent Gatto
}
\references{
The \code{mzTab} specification document and example
files: \url{http://code.google.com/p/mztab/}.
}
\seealso{
Functions to generate metadata (\code{\link{makeMTD}}),
peptide data (\code{\link{makePEP}}) and proteins
(\code{\link{makePRT}}). \code{\link{readMzTabData}} to create
\code{"\linkS4class{MSnSet}"} instances from an \code{mzTab} file.
}

