% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWriteMzTab.R
\name{writeMzTabData}
\alias{writeMzTabData}
\title{Export an MzTab object as mzTab file.}
\usage{
writeMzTabData(
  object,
  file,
  what = c("MT", "PEP", "PRT", "PSM", "SML", "SMF", "SME")
)
}
\arguments{
\item{object}{\link{MzTab} object, either read in by \code{MzTab()} or assembled.}

\item{file}{\code{character(1)} with the file name.}

\item{what}{\code{character} with names of the sections to be written
out. Expected sections are \code{"MT"}, \code{"PEP"}, \code{"PRT"}, \code{"PSM"},
\code{"SML"}, \code{"SMF"}, or \code{"SME"}.}
}
\description{
\code{writeMzTabData} exports an \link{MzTab} object as mzTab file.  Note
that the comment section "COM" are not written out.
}
\author{
Steffen Neumann
}
